backend/requirements.txt
fastapi 
uvicorn 
streamlit 
requests 
pydantic 
python-dotenv 
huggingface-hub 
groq 
phidata
